made controller almost generic ready and fixed CRUD methods with serde

+++ src/controllers/user_controller.rs
@@ -5,7 +5,7 @@ use axum::extract::Path;
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Json};
 use futures::stream::TryStreamExt;
-use mongodb::bson::Document;
+use mongodb::bson::ser;
 use mongodb::bson::{doc, oid::ObjectId};
 use serde_json::{json, Value};
 use std::vec::Vec;
@@ -65,7 +65,11 @@ pub async fn login(Json(payload): Json<Value>) -> impl IntoResponse {
 }
 
 pub async fn get_by_id(Path(id): Path<String>) -> impl IntoResponse {
-    let filter = doc! {"_id": ObjectId::parse_str(id).unwrap()};
+    let mongo_id = ObjectId::parse_str(id);
+    if mongo_id.is_err() {
+        return (StatusCode::NOT_FOUND, Json(json!({"error": "Not found"})));
+    }
+    let filter = doc! {"_id": mongo_id.unwrap()};
     let user = users_coll().await.find_one(filter, None).await.unwrap();
     let json = Json(serde_json::to_value(&user).unwrap());
     match &user {
@@ -86,34 +90,38 @@ pub async fn get_all() -> impl IntoResponse {
 }
 
 pub async fn update(Path(id): Path<String>, Json(payload): Json<Value>) -> impl IntoResponse {
-    let filter = doc! {"_id": ObjectId::parse_str(id).unwrap()};
-    // We convert the json payload into a Document so we can pass it in the update_one function.
-    let mut changes_doc: Document = Document::new();
-    let changes = payload.as_object().unwrap();
-    for (k, v) in changes {
-        if v.is_number() {
-            changes_doc.insert(k, v.as_i64());
-        } else {
-            changes_doc.insert(k, v.as_str());
+    let mongo_id = ObjectId::parse_str(id);
+    if mongo_id.is_err() {
+        return (StatusCode::NOT_FOUND, Json(json!({"error": "Not found"})));
+    }
+    let filter = doc! {"_id": mongo_id.unwrap()};
+    // remove if we make generic controller
+    let doc = ser::to_bson(&payload);
+    match doc {
+        Ok(d) => {
+            let doc = d.as_document().unwrap();
+            let result = users_coll()
+                .await
+                .update_one(filter, doc! {"$set":doc}, None)
+                .await
+                .ok();
+            return (StatusCode::OK, Json(json!({ "status": result })));
+        }
+        Err(e) => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": e.to_string()})),
+            )
         }
     }
-    let changes = doc! {"$set": changes_doc};
-    let result = users_coll()
-        .await
-        .update_one(filter, changes, None)
-        .await
-        .ok();
-    // The result contains the value matchedCount which shows how many values got changed,
-    // by changing this type to JSON we can check if the value is higher than 0, otherwise it's a 404.
-    let result_as_json = Json(serde_json::to_value(&result).unwrap());
-    if result_as_json["matchedCount"].as_i64().unwrap() == 0 {
-        return (StatusCode::NOT_FOUND, Json(json!({"status": "not_found"})));
-    }
-    return (StatusCode::OK, Json(json!({ "status": result })));
 }
 
 pub async fn remove(Path(id): Path<String>) -> impl IntoResponse {
-    let filter = doc! {"_id": ObjectId::parse_str(id).unwrap()};
+    let mongo_id = ObjectId::parse_str(id);
+    if mongo_id.is_err() {
+        return (StatusCode::NOT_FOUND, Json(json!({"error": "Not found"})));
+    }
+    let filter = doc! {"_id": mongo_id.unwrap()};
     let result = users_coll().await.delete_one(filter, None).await.ok();
     // this is similar as what was commented on update, we change the result to json so we can grab the value deletedCount
     // if it didn't delete anything we give a status of not found.
